requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-groq
langchain-huggingface
langchain-graph-retriever
sentence-transformers
faiss-cpu
huggingface-hub
python-dotenv
pydantic
requests
youtube-transcript-api
python-multipart
pypdf